Weapon - laser
Added a new type of weapon - laser. During the game, the energy scale is accumulated. When it is full, the player can fire the laser. The laser travels in a continuous stream from the player to the edge of the screen, dealing damage to all enemies it comes into contact with every frame. Action time 10 seconds.